firmware: move cynthion-specific ladybug Bit constants into moondancer crate
